requirements.txt
# This file is expected by Heroku.

-r requirements/production.txt

django-filter~=23.5
django-ckeditor~=6.7.1
pytest~=7.4.3
django-debug-toolbar~=4.2.0
